% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFieldID2data.R
\name{addFieldID2data}
\alias{addFieldID2data}
\title{add field name and id to a dataframe}
\usage{
addFieldID2data(data, expe = NULL)
}
\arguments{
\item{data}{a dataframe of data}

\item{expe}{name of the experiment}
}
\value{
a dataframe with new cols name_of_experiment,field_name and field_if
}
\description{
add field name and id to a dataframe
}
