% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDataType.R
\name{convertDataType}
\alias{convertDataType}
\title{Conversion data frame text selon type de variables}
\usage{
convertDataType(data)
}
\arguments{
\item{data}{data frame with standardized col to convert}
}
\description{
Conversion data frame text selon type de variables
}
