% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTables.R
\name{dataTables}
\alias{dataTables}
\title{get the list of all data table in a repo}
\usage{
dataTables(expes = "all", noreturn = FALSE)
}
\arguments{
\item{expes}{vector of names of the experiment}

\item{noreturn}{if true the function do not return anything, but experiments are stored in the working environnement}
}
\description{
get the list of all data table in a repo
}
