% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{dictionaries}
\alias{dictionaries}
\title{extract information of data dictionary in the repo}
\usage{
dictionaries(noreturn = FALSE)
}
\arguments{
\item{noreturn}{if true the function do not return anything, but dictionary are stored in the working environnement}
}
\value{
a dataframe with dictionaries found in the repo
}
\description{
extract information of data dictionary in the repo
}
