% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{experiments}
\alias{experiments}
\title{extract information of experiments in the repo}
\usage{
experiments(noreturn = FALSE)
}
\arguments{
\item{noreturn}{if true the function do not return anything, but experiments are stored in the working environnement}
}
\value{
a dataframe with experiments found in the repo
}
\description{
extract information of experiments in the repo
}
