% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractDataExperiment.R
\name{extractDataExperiment}
\alias{extractDataExperiment}
\title{extract data from experiment}
\usage{
extractDataExperiment(expe, variables, addFieldName = T, addExpeName = T)
}
\arguments{
\item{expe}{name of the experiment}

\item{variables}{vector of the standardized names of variables to extract}

\item{addFieldName}{true or false if you want to add field name to the dataframe}

\item{addExpeName}{true or false if you want to add expe name to the dataframe}
}
\value{
an unique dataframe that combine all datasets where the variables to extrat where found
}
\description{
extract data from experiment
}
