% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractExpeData.R
\name{extractExpeData}
\alias{extractExpeData}
\title{Extraction of data from an experimentation based on variables}
\usage{
extractExpeData(expe, variables, addExpeName = T)
}
\arguments{
\item{expe}{name of the experiment}

\item{variables}{vector of the standardized names of variables to extract}

\item{addExpeName}{true or false if you want to add expe name to the final dataframe}
}
\value{
n unique dataframe that combine all datasets where the variables to extrat where found
}
\description{
Extraction of data from an experimentation based on variables
}
