% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{fields}
\alias{fields}
\title{extract information of fields in the repo}
\usage{
fields(noreturn = FALSE)
}
\arguments{
\item{noreturn}{if true the function do not return anything, but fields are stored in the working environnement}
}
\value{
a dataframe with fields found in the repo
}
\description{
extract information of fields in the repo
}
