% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllVariables.R
\name{getAllVariables}
\alias{getAllVariables}
\title{get all variables in a repo in both data and metadata}
\usage{
getAllVariables(expes = "all")
}
\arguments{
\item{expes}{vector of names of experiments, default value all to scan all experiments in the repo}
}
\value{
a named list by experiment with metadata and variables
}
\description{
get all variables in a repo in both data and metadata
}
