% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soils.R
\name{soils}
\alias{soils}
\title{extract information of data dictionary in the repo}
\usage{
soils(noreturn = FALSE)
}
\arguments{
\item{noreturn}{if true the function do not return anything, but soils are stored in the working environnement}
}
\value{
a dataframe with soils metadat found in the repo
}
\description{
extract information of data dictionary in the repo
}
