% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeData.R
\name{summarizeData}
\alias{summarizeData}
\title{Title}
\usage{
summarizeData(data2sum, variables, groups)
}
\arguments{
\item{data2sum}{a dataframe with variables to summarize}

\item{variables}{a vector of numeric variables to summarize with mean, sd, count}

\item{groups}{a vector of grouping variables}
}
\description{
Title
}
