% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{variables}
\alias{variables}
\title{Variables}
\usage{
variables(expes = "all")
}
\arguments{
\item{expes}{vector of names of experiments, default value all to scan all experiments in the repo}
}
\value{
a dataframe with all variables for each experiment
}
\description{
Variables
}
