% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{variables}
\alias{variables}
\title{Title}
\usage{
variables(expes = "all")
}
\arguments{
\item{expes}{un vecteur des noms des expes, par defaut all pour extraire les variables de toutes les expes existantes}
}
\value{
a dataframe with all variables for each experiment
}
\description{
Title
}
